Update data_utils.py
fixed bug in generate performance. normalize_pose has to be applied after the sequence has been gererated, because the model should be feeded with data similar to what  it has been trained with. otherwise it will do strange things. Also, the retransformation that is applied when traj=false has to be done on shape (17,3). also it has to be done after the sequence has been generated. same reason as above.